LayoutTests/media/modern-media-controls/tracks-support/tracks-support-show-panel-fullscreen.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<body>
<video src="../../content/CC+Subtitles.mov" style="width: 320px; height: 240px;" controls autoplay></video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Showing the tracks panel in fullscreen.");

const media = document.querySelector("video");
const shadowRoot = window.internals.shadowRoot(media);

media.addEventListener("webkitbeginfullscreen", () => {
    debug("webkitbeginfullscreen");
});

media.addEventListener("webkitfullscreenchange", () => {
    if (media.webkitDisplayingFullscreen) {
        window.requestAnimationFrame(() => {
            clickOnButton("tracks");
            window.requestAnimationFrame(() => {
                window.requestAnimationFrame(() => {
                    shouldBeEqualToString("shadowRoot.querySelector('.tracks-panel').style.right", "217px");
                    shouldBeEqualToString("shadowRoot.querySelector('.tracks-panel').style.bottom", "101px");

                    debug("");
                    media.remove();
                    finishJSTest();
                });
            });
        });
    }
});

media.addEventListener("play", () => {
    media.pause();
    window.requestAnimationFrame(() => clickOnButton("fullscreen"));
});

function clickOnButton(name)
{
    const bounds = shadowRoot.querySelector(`button.${name}`).getBoundingClientRect();
    eventSender.mouseMoveTo(bounds.left + 1, bounds.top + 1);
    eventSender.mouseDown();
    eventSender.mouseUp();
}

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
